pyproject.toml
[project]
name = "ca-policy-manager"
version = "1.0.0"
description = "Azure Conditional Access Policy-as-Code Management"
authors = [{name = "IAM Team", email = "iam@example.com"}]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "azure-identity>=1.15.0",
    "msgraph-sdk>=1.3.0",
    "click>=8.1.7",
    "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0"
]

[project.scripts]
ca-manager = "ca_manager.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "PL"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
